--- components/BlogPage/BlogPageBody/index.ts
@@ -1,3 +1,0 @@
-import BlogPageBody from "./BlogPageBody";
-
-export default BlogPageBody;
