--- components/BlogPage/BlogPageHeader/BlogPageHeader.tsx
@@ -1,11 +1,0 @@
-const BlogPageHeader = () => {
-  return (
-    <div className="flex justify-center items-center h-10">
-      <h1 className="text-red text-center text-xl md:text-2xl lg:text-3xl xl:text-4xl font-semibold">
-        Blog
-      </h1>
-    </div>
-  );
-};
-
-export default BlogPageHeader;
