--- components/BlogPage/BlogPageHeader/index.ts
@@ -1,3 +1,0 @@
-import BlogPageHeader from "./BlogPageHeader";
-
-export default BlogPageHeader;
