+++ components/PageContent/index.ts
@@ -1,0 +1,3 @@
+import PageContent from "./PageContent";
+
+export default PageContent;
