--- components/ReviewsPage/ReviewsPageBody/index.ts
@@ -1,3 +1,0 @@
-import ReviewsPageBody from "./ReviewsPageBody";
-
-export default ReviewsPageBody;
