--- components/ReviewsPage/ReviewsPageHeader/ReviewsPageHeader.tsx
@@ -1,11 +1,0 @@
-const ReviewsPageHeader = () => {
-  return (
-    <div className="flex justify-center items-center h-10">
-      <h1 className="text-red text-center text-xl md:text-2xl lg:text-3xl xl:text-4xl font-semibold">
-        Testimonials
-      </h1>
-    </div>
-  );
-};
-
-export default ReviewsPageHeader;
