--- components/ReviewsPage/ReviewsPageHeader/index.ts
@@ -1,3 +1,0 @@
-import ReviewsPageHeader from "./ReviewsPageHeader";
-
-export default ReviewsPageHeader;
